requirements.txt
requests==2.32.3
Flask==3.1.0
beautifulsoup4==4.12.3
gunicorn==23.0.0
flask-jwt-extended==4.7.1
python-dotenv==1.0.1